Added some error handling
Created a class to store the error status code and then error message, a middlware is also created to send the error response. The error middlware gets called with an anonymous object of the error class which will act as the first parameter. There, an error response will be sent with the details from the object created on the fly